chore: sweep old bench crumbs
