+++ lib/bench/Cargo.toml
@@ -58,12 +58,11 @@ tokenizers = { version = "0.21.4", default-features = false, features = [
 dynamo-tokens = { workspace = true }
 dynamo-mocker = { workspace = true }
 bytemuck = "1.22"
+dynamo-kv-router = { workspace = true, features = ["bench", "metrics", "shard-metrics"] }
 
 [dev-dependencies]
-dynamo-kv-router = { workspace = true, features = ["bench", "metrics", "shard-metrics"] }
 dynamo-tokens = { workspace = true }
 minstant = "0.1.7"
-plotters = { version = "0.3", default-features = false, features = ["svg_backend", "line_series", "point_series", "full_palette"] }
 tempfile = { workspace = true }
 tokio = { workspace = true, features = ["rt", "macros", "time"] }
 tokio-util = { workspace = true }
