+++ lib/bench/kv_router/common/mod.rs
@@ -1,29 +1,9 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-#![allow(dead_code, unused_imports)]
-
 pub mod args;
-pub mod mooncake_trace;
 pub mod progress;
 pub mod replay;
 pub mod results;
 pub mod sweep;
 pub mod trace_gen;
-
-pub use args::{CommonArgs, init_sequence_logging};
-pub use mooncake_trace::{
-    MooncakeRequest, duplicate_traces, expand_trace_lengths, load_mooncake_trace,
-    local_block_hash_from_id, partition_trace, process_mooncake_trace, scale_mooncake_trace,
-    tokens_from_request,
-};
-pub use progress::make_progress_bar;
-pub use replay::{
-    NoopSequencePublisher, WorkerReplayArtifacts, default_mock_engine_args,
-    generate_replay_artifacts, maybe_rescale_ready_span,
-};
-pub use results::{
-    BenchmarkResults, BenchmarkRun, PercentileStats, compute_benchmark_run,
-    print_benchmark_results_percentiles,
-};
-pub use sweep::{compute_sweep_durations, median, plot_sweep, print_sweep_summary};
